# misc/API/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Fast Food FIAP",
    "description": "Projeto de um sistema para lanchonete realizado para a Pós-Graduação de Arquitetura de Sistemas da FIAP",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://${local.load_balancer_dns}",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/swagger-ui/index": {
      "get": {
        "operationId": "Swagger",
        "responses": {
          "200": {
            "description": "200 response",
            "headers": {
              "Access-Control-Allow-Origin": {
                "style": "simple",
                "explode": false,
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {}
          },
          "default": {
            "description": "",
            "content": {}
          }
        },
        "x-amazon-apigateway-integration": {
          "httpMethod": "GET",
          "uri": "http://${local.load_balancer_dns}/swagger-ui/index.html",
          "responses": {
            "default": {
              "statusCode": "200",
              "responseParameters": {
                "method.response.header.Access-Control-Allow-Origin": "'*'"
              }
            }
          },
          "passthroughBehavior": "when_no_match",
          "type": "http"
        }
      }
    },
...
    "/orders": {
      "get": {
        "tags": [
          "order-controller"
        ],
        "operationId": "getOrders",
        "parameters": [
          {
            "name": "cpf_cliente",
            "in": "header",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "senha_cliente",
            "in": "header",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetOrderResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionDetails"
                }
              }
            }
          },
          "default": {
            "description": "",
            "content": {}
          }
        },
        "security": [
          {
            "lambda_authorizer_cpf": []
          }
        ],
        "x-amazon-apigateway-integration": {
          "httpMethod": "GET",
          "payloadFormatVersion": "1.0",
          "requestParameters": {
            "integration.request.header.microsservice": "'ms_pedido'"
          },
          "type": "HTTP_PROXY",
          "uri": "http://${local.load_balancer_dns}/orders"
        }
      },
      "post": {
        "tags": [
          "order-controller"
        ],
        "operationId": "createOrder",
        "parameters": [
          {
            "name": "cpf_cliente",
            "in": "header",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "senha_cliente",
            "in": "header",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateOrderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionDetails"
                }
              }
            }
          },
          "default": {
            "description": "",
            "content": {}
          }
        },
        "security": [
          {
            "lambda_authorizer_cpf": []
          }
        ],
        "x-amazon-apigateway-integration": {
          "httpMethod": "POST",
          "payloadFormatVersion": "1.0",
          "requestParameters": {
            "integration.request.header.microsservice": "'ms_orquestrador'"
          },
          "type": "HTTP_PROXY",
          "uri": "http://${local.load_balancer_dns}/orders"
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CustomError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "field": {
            "type": "string"
          },
          "attemptedValue": {
            "type": "object"
          }
        }
      },
      "ExceptionDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomError"
            }
          }
        }
      },
...
      "CreateOrderRequest": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Item"
            }
          }
        }
      },
      "Item": {
        "type": "object",
        "properties": {
          "idProduct": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "itemValue": {
            "type": "number"
          },
          "totalItemValue": {
            "type": "number"
          }
        }
      },
      "CreateOrderResponse": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Item"
            }
          }
        }
      },
...
      "GetOrderResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Item"
            }
          },
          "totalValue": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "READY",
              "IN_PREPARATION",
              "RECEIVED",
              "COMPLETED"
            ]
          },
          "paymentStatus": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "lambda_authorizer_cpf": {
        "type": "apiKey",
        "name": "auth",
        "in": "header",
        "x-amazon-apigateway-authtype": "custom",
        "x-amazon-apigateway-authorizer": {
          "type": "request",
          "identitySource": "method.request.header.cpf_cliente, method.request.header.senha_cliente",
          "authorizerCredentials": "lambda_arn",
          "authorizerUri": "lambda_uri",
          "authorizerResultTtlInSeconds": 0
        }
      }
    }
  }
}